Use int type instead of size_t in hello world
Program now terminates in reasonable time (probably an overflow issue
before)

+++ 1lab/intro/0.2_hello_world/hello.cpp
@@ -1,8 +1,8 @@
 #include "hello.h"
 
-void hello(std::ostream& out, const std::string& helloee, const size_t times) {
+void hello(std::ostream& out, const std::string& helloee, const int times) {
   out << "Hello,";
-  for(size_t i = 0; i < times; ++i) {
+  for(int i = 0; i < times; ++i) {
     out << " " << helloee;
   }
   out << "!";
