+++ 1lab/intro/0.2_hello_world/hello.h
@@ -2,4 +2,4 @@
 #include <iostream>
 #include<string>
 
-void hello(std::ostream& out, const std::string& helloee, const size_t times);
+void hello(std::ostream& out, const std::string& helloee, const int times);
